+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 middlemac-extras change log
 ===========================
 
+- Version 1.0.6 / 2016-April-08
+
+  - If `middleman-targets` is also installed, support images without extensions
+    for it, too. Make sure to active it first, though.
+    
+  - Added CLI topic to help file.
+
 - Version 1.0.5 / 2016-April-08
 
   - Fixed the implementation of css_image_sizes so that full path is not